requirements.txt
python-dotenv
openai
slack_sdk
slack_bolt
lyzr_agent_api
requests